Creation of data transfer protocol and correction of many errors
The objects of the pollfd structure are stored in a vector instead of an array.
Many for loops have been replaced with for_each.

In the client, the release of the mutex is fixed, which was not released during errors.
Fix bugs when one of the clients exits and others stop working or the server is down.
The waiting for the registration server for a new client has been removed, which allows other registered users to continue communicating.
Fixed checking of customer name uniqueness.
And also fixes of many small bugs.

Added the ability to enter your server work address, as well as the server timeout.
Added the ability to enter your address to connect to the server.
Added data transfer protocol.